src/requirements.txt
numpy>=1.26.4
pandas>=2.2.0
python-binance>=1.0.19